+++ src/m365/teams/commands/channel/channel-get.spec.ts
@@ -13,7 +13,7 @@ describe(commands.CHANNEL_GET, () => {
   let log: string[];
   let logger: Logger;
   let loggerLogSpy: sinon.SinonSpy;
-  
+
   before(() => {
     sinon.stub(auth, 'restoreAuth').callsFake(() => Promise.resolve());
     sinon.stub(appInsights, 'trackEvent').callsFake(() => { });
@@ -207,7 +207,7 @@ describe(commands.CHANNEL_GET, () => {
 
   it('fails to get team when team does not exists', (done) => {
     sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf(`/beta/groups?$filter=resourceProvisioningOptions/Any(x:x eq 'Team') and displayName eq '`) > -1) {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
         return Promise.resolve({ value: [] });
       }
       return Promise.reject('The specified team does not exist in the Microsoft Teams');
@@ -231,9 +231,60 @@ describe(commands.CHANNEL_GET, () => {
     });
   });
 
+  it('fails when team name does not exist', (done) => {
+    sinon.stub(request, 'get').callsFake((opts) => {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
+        return Promise.resolve({
+          "@odata.context": "https://graph.microsoft.com/v1.0/$metadata#teams",
+          "@odata.count": 1,
+          "value": [
+            {
+              "id": "00000000-0000-0000-0000-000000000000",
+              "createdDateTime": null,
+              "displayName": "Team Name",
+              "description": "Team Description",
+              "internalId": null,
+              "classification": null,
+              "specialization": null,
+              "visibility": null,
+              "webUrl": null,
+              "isArchived": false,
+              "isMembershipLimitedToOwners": null,
+              "memberSettings": null,
+              "guestSettings": null,
+              "messagingSettings": null,
+              "funSettings": null,
+              "discoverySettings": null,
+              "resourceProvisioningOptions": []
+            }
+          ]
+        }
+        );
+      }
+      return Promise.reject('Invalid request');
+    });
+
+    command.action(logger, {
+      options: {
+        debug: true,
+        teamName: 'Team Name',
+        channelName: 'Channel Name',
+        tabName: 'Tab Name'
+      }
+    }, (err?: any) => {
+      try {
+        assert.strictEqual(JSON.stringify(err), JSON.stringify(new CommandError(`The specified team does not exist in the Microsoft Teams`)));
+        done();
+      }
+      catch (e) {
+        done(e);
+      }
+    });
+  });
+
   it('fails when multiple teams with same name exists', (done) => {
     sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf(`/beta/groups?$filter=resourceProvisioningOptions/Any(x:x eq 'Team') and displayName eq '`) > -1) {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
         return Promise.resolve({
           "@odata.context": "https://graph.microsoft.com/v1.0/$metadata#teams",
           "@odata.count": 2,
@@ -254,7 +305,8 @@ describe(commands.CHANNEL_GET, () => {
               "guestSettings": null,
               "messagingSettings": null,
               "funSettings": null,
-              "discoverySettings": null
+              "discoverySettings": null,
+              "resourceProvisioningOptions": ["Team"]
             },
             {
               "id": "00000000-0000-0000-0000-000000000000",
@@ -272,7 +324,8 @@ describe(commands.CHANNEL_GET, () => {
               "guestSettings": null,
               "messagingSettings": null,
               "funSettings": null,
-              "discoverySettings": null
+              "discoverySettings": null,
+              "resourceProvisioningOptions": ["Team"]
             }
           ]
         }
@@ -362,7 +415,7 @@ describe(commands.CHANNEL_GET, () => {
 
   it('should get channel information for the Microsoft Teams team by name', (done) => {
     sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf(`/beta/groups?$filter=resourceProvisioningOptions/Any(x:x eq 'Team') and displayName eq '`) > -1) {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
         return Promise.resolve({
           "value": [
             {
@@ -381,7 +434,8 @@ describe(commands.CHANNEL_GET, () => {
               "guestSettings": null,
               "messagingSettings": null,
               "funSettings": null,
-              "discoverySettings": null
+              "discoverySettings": null,
+              "resourceProvisioningOptions": ["Team"]
             }
           ]
         });